+++ ocbpy/tests/test_pysat.py
@@ -26,47 +26,6 @@ except ImportError:
     no_pysat = True
 
 
-def add_mag_coords(inst, dat='dummy2', lat='latitude', lon='longitude',
-                   alt='altitude'):
-    """Add magnetic coordinates to a pysat Instrument.
-
-    Parameters
-    ----------
-    inst : pysat.Instrument
-        pysat Instrument object with data to be modified
-    dat : str
-        pysat data variable with the desired output dimensions
-    lat : str
-        pysat data variable or coordinate for the latitude
-    lon : str
-        pysat data variable or coordinate for the longitude
-    alt : str
-        pysat data variable or coordinate for the altitude
-
-    """
-    # Define the dimensions and make sure they behave as expected
-    dims = [inst.data.dims[dd] for dd in inst[dat].dims]
-    mlat = np.full(shape=dims, fill_value=np.nan)
-    mlt = np.full(shape=dims, fill_value=np.nan)
-
-    # Get the magnetic coordinates
-    lat_in, lon_in = np.meshgrid(inst[lat].values, inst[lon].values)
-    for i, height in enumerate(inst[alt]):
-        for j, lin in enumerate(lat_in):
-            out = aacgmv2.convert_latlon_arr(lin, lon_in[j], height,
-                                             inst.date)
-            out_mlt = aacgmv2.convert_mlt(out[1], inst.date)
-            mlat[:, :, j, i] = np.full(shape=dims[:2], fill_value=out[0])
-            mlt[:, :, j, i] = np.full(shape=dims[:2], fill_value=out_mlt)
-
-    # Assign the magnetic data to the instrument
-    inst.data = inst.data.assign({"mlat": (inst[dat].dims, mlat),
-                                  "mlt": (inst[dat].dims, mlt)})
-    inst.meta['mlat'] = {inst.meta.labels.units: 'degrees'}
-    inst.meta['mlt'] = {inst.meta.labels.units: 'hours'}
-    return
-
-
 @unittest.skipIf(no_pysat, "pysat not installed, cannot test routines")
 class TestPysatUtils(unittest.TestCase):
     """Tests for using ocbpy with pysat data."""
@@ -79,20 +38,18 @@ class TestPysatUtils(unittest.TestCase):
         self.pysat_key = 'dummy1'
         self.pysat_lat = 'latitude'
         self.notes = None
-        self.isvector = False
         self.test_inst = None
         self.ocb = None
         self.added_keys = list()
         self.pysat_keys = list()
-        self.arevectors = list()
         self.del_time = 600
         return
 
     def tearDown(self):
         """Tear down the testing environment."""
-        del self.ocb_key, self.pysat_key, self.notes, self.isvector
+        del self.ocb_key, self.pysat_key, self.notes, self.del_time
         del self.test_inst, self.ocb, self.added_keys, self.pysat_keys
-        del self.arevectors, self.del_time, self.pysat_lat
+        del self.pysat_lat
         return
 
     def eval_ocb_metadata(self):
@@ -118,7 +75,7 @@ class TestPysatUtils(unittest.TestCase):
                              self.test_inst.meta.labels.fill_val]),
                          ", not np.nan"]))
 
-        if self.pysat_key is not None and not self.isvector:
+        if self.pysat_key is not None:
             # Test the elements that are identical
             for ll in [self.test_inst.meta.labels.units,
                        self.test_inst.meta.labels.min_val,
@@ -128,22 +85,41 @@ class TestPysatUtils(unittest.TestCase):
                     if np.isnan(self.test_inst.meta[self.pysat_key][ll]):
                         self.assertTrue(
                             np.isnan(self.test_inst.meta[self.ocb_key][ll]))
-                    else:
+                    elif ll != self.test_inst.meta.labels.fill_val:
+                        # The OCB fill value is NaN, regardless of prior value
                         self.assertEqual(
                             self.test_inst.meta[self.ocb_key][ll],
-                            self.test_inst.meta[self.pysat_key][ll])
+                            self.test_inst.meta[self.pysat_key][ll],
+                            msg="unequal fill values [{:s} and {:s}]".format(
+                                self.ocb_key, self.pysat_key))
                 except TypeError:
-                    self.assertRegex(self.test_inst.meta[self.ocb_key][ll],
-                                     self.test_inst.meta[self.pysat_key][ll])
+                    ocb_len = len(self.test_inst.meta[self.ocb_key][ll])
+                    pysat_len = len(self.test_inst.meta[self.pysat_key][ll])
+                    if pysat_len == 0:
+                        self.assertGreaterEqual(ocb_len, pysat_len)
+                    else:
+                        self.assertRegex(
+                            self.test_inst.meta[self.ocb_key][ll],
+                            self.test_inst.meta[self.pysat_key][ll],
+                            msg="".join(["Meta label ", ll, ": OCB key ",
+                                         self.ocb_key, " value `",
+                                         self.test_inst.meta[self.ocb_key][ll],
+                                         "` not in pysat key ", self.pysat_key,
+                                         " value `", self.test_inst.meta[
+                                             self.pysat_key][ll], "`"]))
 
         # Test the elements that have "OCB" appended to the text
         sline = self.test_inst.meta[self.ocb_key][
             self.test_inst.meta.labels.name].split(" ")
         self.assertRegex(sline[0], "OCB")
-        if not self.isvector and self.pysat_key is not None:
+        note_line = self.test_inst.meta[self.ocb_key][
+            self.test_inst.meta.labels.notes]
+        if self.pysat_key is not None and note_line.find('scaled using') < 0:
             self.assertRegex(" ".join(sline[1:]),
                              self.test_inst.meta[self.pysat_key][
-                                 self.test_inst.meta.labels.name])
+                                 self.test_inst.meta.labels.name],
+                             msg="Bad long name for {:}; notes are: {:}".format(
+                                 self.pysat_key, note_line))
 
         # Test the remaining elements
         self.assertEqual(self.test_inst.meta[self.ocb_key][
@@ -160,46 +136,35 @@ class TestPysatUtils(unittest.TestCase):
             return
 
         # For a test instrument, evaluate attributes
-        if len(self.arevectors) < len(self.added_keys):
-            self.arevectors = [False for okey in self.added_keys]
-
         for i, self.ocb_key in enumerate(self.added_keys):
             # Test to see that data was added
             self.assertIn(self.ocb_key, self.test_inst.variables)
 
             # Test the metadata
             self.pysat_key = self.pysat_keys[i]
-            self.isvector = self.arevectors[i]
             self.eval_ocb_metadata()
 
             # Test to see that data within the time tolerance of the test OCBs
             # has OCB locations and other data is NaN
-            match_data = np.array(self.test_inst[self.ocb_key])
-            if self.isvector:
-                mask_data = np.not_equal(match_data, None)
-            else:
-                mask_data = np.isfinite(match_data)
+            match_data = self.test_inst[self.ocb_key]
+            mask_data = np.isfinite(match_data)
 
-            self.assertTrue(mask_data.any())
+            self.assertTrue(np.array(mask_data).any())
 
-            mind = np.where(mask_data)
-            match_data = match_data[mind]
-            match_time = self.test_inst.index[mind[0]]
+            mind = np.where(np.array(mask_data))
+            match_data = np.array(match_data)[mind]
+            match_time = np.unique(self.test_inst.index[mind[0]])
 
             for ii in match_time:
-                check_time = abs(ii - self.ocb.dtime).min().total_seconds()
-                self.assertLessEqual(check_time, self.del_time)
+                check_time = abs(pds.to_datetime(ii)
+                                 - self.ocb.dtime).min().total_seconds()
+                self.assertLessEqual(check_time, self.del_time,
+                                     msg="".join(["bad time difference for ",
+                                                  "OCB key ", self.ocb_key]))
 
-            if self.isvector:
-                self.assertTrue(isinstance(match_data[0],
-                                           ocbpy.ocb_scaling.VectorData),
-                                msg='Unexpected class {:} != {:}'.format(
-                                    repr(type(match_data[0])),
-                                    repr(ocbpy.ocb_scaling.VectorData)))
-
-            elif(self.pysat_keys[i] is not None
-                 and self.pysat_keys[i] not in [self.pysat_lat, 'mlt']):
-                pysat_data = self.test_inst[self.pysat_keys[i]].where(mask_data)
+            if(self.pysat_key is not None
+               and self.pysat_key not in [self.pysat_lat, 'mlt']):
+                pysat_data = self.test_inst[self.pysat_key].where(mask_data)
 
                 # Get the scaling radius
                 if hasattr(self.ocb, "r"):
@@ -209,19 +174,20 @@ class TestPysatUtils(unittest.TestCase):
                     rscale = (self.ocb.ocb.r
                               / (90.0 - abs(self.ocb.ocb.boundary_lat)))**2
 
-                # Evaluate the data
-                self.assertGreaterEqual(match_data.min(),
-                                        pysat_data.min() * rscale.min(),
-                                        msg="".join([
-                                            "bad comparison between ",
-                                            self.ocb_key, " and ",
-                                            self.pysat_keys[i]]))
-                self.assertGreaterEqual(pysat_data.max() * rscale.max(),
-                                        match_data.max(),
-                                        msg="".join([
-                                            "bad comparison between ",
-                                            self.ocb_key, " and ",
-                                            self.pysat_keys[i]]))
+                # Evaluate the non-vector data
+                if self.ocb_key.find(self.pysat_key) >= 0:
+                    self.assertGreaterEqual(match_data.min(),
+                                            pysat_data.min() * rscale.min(),
+                                            msg="".join([
+                                                "bad comparison between ",
+                                                self.ocb_key, " and ",
+                                                self.pysat_key]))
+                    self.assertGreaterEqual(pysat_data.max() * rscale.max(),
+                                            match_data.max(),
+                                            msg="".join([
+                                                "bad comparison between ",
+                                                self.ocb_key, " and ",
+                                                self.pysat_key]))
 
         return
 
@@ -280,10 +246,8 @@ class TestPysatMethods(TestPysatUtils):
         self.pysat_key = 'dummy1'
         self.pysat_lat = 'latitude'
         self.notes = None
-        self.isvector = False
         self.added_keys = list()
         self.pysat_keys = list()
-        self.arevectors = list()
         self.del_time = 600
 
         # Set the method default values
@@ -321,9 +285,9 @@ class TestPysatMethods(TestPysatUtils):
     def tearDown(self):
         """Tear down after each test."""
 
-        del self.ocb_key, self.pysat_key, self.notes, self.isvector
+        del self.ocb_key, self.pysat_key, self.notes
         del self.test_inst, self.ocb, self.added_keys, self.pysat_keys
-        del self.arevectors, self.test_file, self.log_capture, self.pysat_lat
+        del self.test_file, self.log_capture, self.pysat_lat
         del self.lout, self.lwarn, self.ocb_kw, self.pysat_var2, self.del_time
         return
 
@@ -338,11 +302,11 @@ class TestPysatMethods(TestPysatUtils):
     def test_add_ocb_to_metadata_vector(self):
         """Test the metadata adding routine for vector data."""
         ocb_pysat.add_ocb_to_metadata(self.test_inst, "ocb_test",
-                                      self.pysat_key, notes="test notes",
+                                      self.pysat_key,
+                                      notes="test notes scaled using None",
                                       isvector=True)
 
-        self.notes = 'test notes'
-        self.isvector = True
+        self.notes = 'test notes scaled using None'
         self.eval_ocb_metadata()
 
     def test_no_overwrite_metadata(self):
@@ -406,7 +370,6 @@ class TestPysatMethods(TestPysatUtils):
         self.pysat_keys = [aa.split("_ocb")[0] for aa in self.added_keys]
         self.assertIn('r_corr', self.pysat_keys)
         self.pysat_keys[self.pysat_keys.index("r_corr")] = None
-        self.arevectors = [False for i in self.added_keys]
         self.test_ocb_added()
         return
 
@@ -422,10 +385,15 @@ class TestPysatMethods(TestPysatUtils):
         new_lat[new_lat > 0] *= -1.0
         imax = np.where(np.nanmax(new_lat) == new_lat)  # Needed for model data
         new_lat[imax] = 0.0
-        try:
-            self.test_inst[self.pysat_lat] = new_lat
-        except ValueError:
-            self.test_inst[self.pysat_lat].values = new_lat
+        if(not self.test_inst.pandas_format
+           and self.pysat_lat in self.test_inst.data.coords):
+            self.test_inst.data = self.test_inst.data.assign_coords(
+                {self.pysat_lat: new_lat})
+        else:
+            try:
+                self.test_inst[self.pysat_lat] = new_lat
+            except ValueError:
+                self.test_inst[self.pysat_lat].values = new_lat
 
         # Add the OCB data to the Instrument and evaluate the output
         ocb_pysat.add_ocb_to_data(self.test_inst, self.pysat_lat, "mlt",
@@ -478,15 +446,12 @@ class TestPysatMethods(TestPysatUtils):
         self.added_keys = [kk for kk in self.test_inst.meta.keys()
                            if kk.find('_ocb') > 0]
         self.pysat_keys = list()
-        self.arevectors = list()
         for aa in self.added_keys:
             pp = aa.split("_ocb")[0]
-            self.arevectors.append(False)
             if pp == "r_corr":
                 pp = None
             elif pp not in self.test_inst.variables:
                 pp = self.pysat_key
-                self.arevectors[-1] = True
             self.pysat_keys.append(pp)
 
         self.test_ocb_added()
@@ -507,15 +472,12 @@ class TestPysatMethods(TestPysatUtils):
         self.added_keys = [kk for kk in self.test_inst.meta.keys()
                            if kk.find('_ocb') > 0]
         self.pysat_keys = list()
-        self.arevectors = list()
         for aa in self.added_keys:
             pp = aa.split("_ocb")[0]
-            self.arevectors.append(False)
             if pp == "r_corr":
                 pp = None
             elif pp not in self.test_inst.variables:
                 pp = self.pysat_key
-                self.arevectors[-1] = True
             self.pysat_keys.append(pp)
 
         self.test_ocb_added()
@@ -537,15 +499,12 @@ class TestPysatMethods(TestPysatUtils):
         self.added_keys = [kk for kk in self.test_inst.meta.keys()
                            if kk.find('_ocb') > 0]
         self.pysat_keys = list()
-        self.arevectors = list()
         for aa in self.added_keys:
             pp = aa.split("_ocb")[0]
-            self.arevectors.append(False)
             if pp == "r_corr":
                 pp = None
             elif pp not in self.test_inst.variables:
                 pp = self.pysat_key
-                self.arevectors[-1] = True
             self.pysat_keys.append(pp)
         self.test_ocb_added()
         return
@@ -567,15 +526,12 @@ class TestPysatMethods(TestPysatUtils):
         self.added_keys = [kk for kk in self.test_inst.meta.keys()
                            if kk.find('_ocb') > 0]
         self.pysat_keys = list()
-        self.arevectors = list()
         for aa in self.added_keys:
             pp = aa.split("_ocb")[0]
-            self.arevectors.append(False)
             if pp == "r_corr":
                 pp = None
             elif pp not in self.test_inst.variables:
                 pp = self.pysat_key
-                self.arevectors[-1] = True
             self.pysat_keys.append(pp)
 
         self.test_ocb_added()
@@ -689,10 +645,8 @@ class TestPysatMethodsEAB(TestPysatMethods):
         self.pysat_key = 'dummy1'
         self.pysat_lat = 'latitude'
         self.notes = None
-        self.isvector = False
         self.added_keys = list()
         self.pysat_keys = list()
-        self.arevectors = list()
         self.del_time = 600
 
         # Set the method defaults
@@ -731,8 +685,8 @@ class TestPysatMethodsEAB(TestPysatMethods):
 
         del self.test_file, self.log_capture, self.ocb, self.test_inst
         del self.lout, self.lwarn, self.ocb_kw, self.pysat_var2, self.ocb_key
-        del self.pysat_key, self.notes, self.isvector, self.pysat_lat
-        del self.added_keys, self.pysat_keys, self.arevectors, self.del_time
+        del self.pysat_key, self.notes, self.pysat_lat, self.del_time
+        del self.added_keys, self.pysat_keys
         return
 
 
@@ -747,10 +701,8 @@ class TestPysatMethodsDual(TestPysatMethods):
         self.pysat_key = 'dummy1'
         self.pysat_lat = 'latitude'
         self.notes = None
-        self.isvector = False
         self.added_keys = list()
         self.pysat_keys = list()
-        self.arevectors = list()
         self.del_time = 600
 
         # Set the method defaults
@@ -793,8 +745,8 @@ class TestPysatMethodsDual(TestPysatMethods):
 
         del self.test_file, self.log_capture, self.ocb, self.test_inst
         del self.lout, self.lwarn, self.ocb_kw, self.pysat_var2, self.pysat_lat
-        del self.ocb_key, self.pysat_key, self.notes, self.isvector
-        del self.added_keys, self.pysat_keys, self.arevectors, self.del_time
+        del self.ocb_key, self.pysat_key, self.notes, self.del_time
+        del self.added_keys, self.pysat_keys
         return
 
 
@@ -810,10 +762,8 @@ class TestPysatMethods2DXarray(TestPysatMethods):
         self.pysat_key = 'dummy1'
         self.pysat_lat = 'latitude'
         self.notes = None
-        self.isvector = False
         self.added_keys = list()
         self.pysat_keys = list()
-        self.arevectors = list()
         self.del_time = 600
 
         # Set the method defaults
@@ -854,8 +804,8 @@ class TestPysatMethods2DXarray(TestPysatMethods):
 
         del self.test_file, self.log_capture, self.ocb, self.test_inst
         del self.lout, self.lwarn, self.ocb_kw, self.pysat_var2, self.pysat_lat
-        del self.ocb_key, self.pysat_key, self.notes, self.isvector
-        del self.added_keys, self.pysat_keys, self.arevectors, self.del_time
+        del self.ocb_key, self.pysat_key, self.notes, self.del_time
+        del self.added_keys, self.pysat_keys
         return
 
 
@@ -870,10 +820,8 @@ class TestPysatMethodsXarray(TestPysatMethods):
         self.pysat_key = 'dummy1'
         self.pysat_lat = 'latitude'
         self.notes = None
-        self.isvector = False
         self.added_keys = list()
         self.pysat_keys = list()
-        self.arevectors = list()
         self.del_time = 600
 
         # Set the method defaults
@@ -914,8 +862,8 @@ class TestPysatMethodsXarray(TestPysatMethods):
 
         del self.test_file, self.log_capture, self.ocb, self.test_inst
         del self.lout, self.lwarn, self.ocb_kw, self.pysat_var2, self.del_time
-        del self.ocb_key, self.pysat_key, self.notes, self.isvector
-        del self.added_keys, self.pysat_keys, self.arevectors, self.pysat_lat
+        del self.ocb_key, self.pysat_key, self.notes, self.pysat_lat
+        del self.added_keys, self.pysat_keys
         return
 
 
@@ -928,12 +876,10 @@ class TestPysatMethodsModel(TestPysatMethods):
         # Set the utility defaults
         self.ocb_key = 'ocb_test'
         self.pysat_key = 'dummy2'
-        self.pysat_lat = 'mlat'
+        self.pysat_lat = 'latitude'
         self.notes = None
-        self.isvector = False
         self.added_keys = list()
         self.pysat_keys = list()
-        self.arevectors = list()
         self.del_time = 600
 
         # Set the method defaults
@@ -948,7 +894,6 @@ class TestPysatMethodsModel(TestPysatMethods):
 
         self.pysat_var2 = 'dummy2'
         self.test_inst = pysat.Instrument('pysat', 'testmodel',
-                                          num_samples=50400,
                                           clean_level='clean',
                                           update_files=True,
                                           file_date_range=pds.date_range(
@@ -959,9 +904,6 @@ class TestPysatMethodsModel(TestPysatMethods):
         load_kwargs = {'date': self.ocb.dtime[self.ocb.rec_ind]}
         if version.Version(pysat.__version__) > version.Version('3.0.1'):
             load_kwargs['use_header'] = True
-
-        self.test_inst.custom_attach(add_mag_coords,
-                                     kwargs={"dat": self.pysat_var2})
         self.test_inst.load(**load_kwargs)
 
         self.lwarn = u""
@@ -977,8 +919,23 @@ class TestPysatMethodsModel(TestPysatMethods):
 
         del self.test_file, self.log_capture, self.ocb, self.test_inst
         del self.lout, self.lwarn, self.ocb_kw, self.pysat_var2, self.del_time
-        del self.ocb_key, self.pysat_key, self.notes, self.isvector
-        del self.added_keys, self.pysat_keys, self.arevectors, self.pysat_lat
+        del self.ocb_key, self.pysat_key, self.notes, self.pysat_lat
+        del self.added_keys, self.pysat_keys
+        return
+
+    def test_mismatched_vector_data(self):
+        """Test that vector data with different dimensions fails."""
+        with self.assertRaisesRegex(ValueError,
+                                    'vector variables must all have the same'):
+            ocb_pysat.add_ocb_to_data(self.test_inst, self.pysat_lat, "mlt",
+                                      evar_names=['vect_evar'],
+                                      vector_names={
+                                          'vect_evar':
+                                          {'aacgm_n': self.pysat_key,
+                                           'aacgm_e': self.pysat_lat,
+                                           'dat_name': 'vect',
+                                           'dat_units': 'm/s'}},
+                                      ocb=self.ocb, max_sdiff=self.del_time)
         return
 
 
@@ -995,10 +952,8 @@ class TestPysatCustMethods(TestPysatUtils):
         self.pysat_key = 'dummy1'
         self.pysat_lat = 'latitude'
         self.notes = None
-        self.isvector = False
         self.added_keys = list()
         self.pysat_keys = list()
-        self.arevectors = list()
         self.del_time = 600
 
         # Set the method defaults
@@ -1030,11 +985,10 @@ class TestPysatCustMethods(TestPysatUtils):
     def tearDown(self):
         """Clean the test environment."""
 
-        del self.ocb_key, self.pysat_key, self.notes, self.isvector
-        del self.added_keys, self.pysat_keys, self.arevectors, self.pysat_lat
-        del self.test_file, self.log_capture, self.test_inst, self.ocb
-        del self.lwarn, self.lout, self.cust_kwargs, self.pysat_var2
-        del self.del_time
+        del self.ocb_key, self.pysat_key, self.notes, self.pysat_lat
+        del self.added_keys, self.pysat_keys, self.test_inst, self.ocb
+        del self.test_file, self.log_capture, self.cust_kwargs, self.pysat_var2
+        del self.lwarn, self.lout, self.del_time
         return
 
     def test_load(self):
@@ -1136,15 +1090,12 @@ class TestPysatCustMethods(TestPysatUtils):
         self.added_keys = [kk for kk in self.test_inst.meta.keys()
                            if kk.find('_ocb') > 0]
         self.pysat_keys = list()
-        self.arevectors = list()
         for aa in self.added_keys:
             pp = aa.split("_ocb")[0]
-            self.arevectors.append(False)
             if pp == "r_corr":
                 pp = None
             elif pp not in self.test_inst.variables:
                 pp = self.pysat_key
-                self.arevectors[-1] = True
             self.pysat_keys.append(pp)
         self.test_ocb_added()
         return
@@ -1166,15 +1117,12 @@ class TestPysatCustMethods(TestPysatUtils):
         self.added_keys = [kk for kk in self.test_inst.meta.keys()
                            if kk.find('_ocb') > 0]
         self.pysat_keys = list()
-        self.arevectors = list()
         for aa in self.added_keys:
             pp = aa.split("_ocb")[0]
-            self.arevectors.append(False)
             if pp == "r_corr":
                 pp = None
             elif pp not in self.test_inst.variables:
                 pp = self.pysat_key
-                self.arevectors[-1] = True
             self.pysat_keys.append(pp)
         self.test_ocb_added()
         return
@@ -1194,15 +1142,12 @@ class TestPysatCustMethods(TestPysatUtils):
         self.added_keys = [kk for kk in self.test_inst.meta.keys()
                            if kk.find('_ocb') > 0]
         self.pysat_keys = list()
-        self.arevectors = list()
         for aa in self.added_keys:
             pp = aa.split("_ocb")[0]
-            self.arevectors.append(False)
             if pp == "r_corr":
                 pp = None
             elif pp not in self.test_inst.variables:
                 pp = self.pysat_key
-                self.arevectors[-1] = True
             self.pysat_keys.append(pp)
 
         self.test_ocb_added()
@@ -1226,15 +1171,12 @@ class TestPysatCustMethods(TestPysatUtils):
         self.added_keys = [kk for kk in self.test_inst.meta.keys()
                            if kk.find('_ocb') > 0]
         self.pysat_keys = list()
-        self.arevectors = list()
         for aa in self.added_keys:
             pp = aa.split("_ocb")[0]
-            self.arevectors.append(False)
             if pp == "r_corr":
                 pp = None
             elif pp not in self.test_inst.variables:
                 pp = self.pysat_key
-                self.arevectors[-1] = True
             self.pysat_keys.append(pp)
 
         self.test_ocb_added()
@@ -1337,10 +1279,8 @@ class TestPysatCustMethodsEAB(TestPysatCustMethods):
         self.pysat_key = 'dummy1'
         self.pysat_lat = 'latitude'
         self.notes = None
-        self.isvector = False
         self.added_keys = list()
         self.pysat_keys = list()
-        self.arevectors = list()
         self.del_time = 600
 
         # Set the method defaults
@@ -1372,11 +1312,10 @@ class TestPysatCustMethodsEAB(TestPysatCustMethods):
     def tearDown(self):
         """Clean the test environment."""
 
-        del self.ocb_key, self.pysat_key, self.notes, self.isvector
-        del self.added_keys, self.pysat_keys, self.arevectors, self.del_time
-        del self.test_file, self.log_capture, self.test_inst, self.ocb
-        del self.lwarn, self.lout, self.cust_kwargs, self.pysat_var2
-        del self.pysat_lat
+        del self.ocb_key, self.pysat_key, self.notes, self.del_time
+        del self.added_keys, self.pysat_keys, self.test_inst, self.ocb
+        del self.test_file, self.log_capture, self.cust_kwargs, self.pysat_var2
+        del self.lwarn, self.lout, self.pysat_lat
         return
 
 
@@ -1392,10 +1331,8 @@ class TestPysatCustMethodsDual(TestPysatCustMethods):
         self.pysat_key = 'dummy1'
         self.pysat_lat = 'latitude'
         self.notes = None
-        self.isvector = False
         self.added_keys = list()
         self.pysat_keys = list()
-        self.arevectors = list()
         self.del_time = 600
 
         # Set the method defaults
@@ -1431,11 +1368,10 @@ class TestPysatCustMethodsDual(TestPysatCustMethods):
     def tearDown(self):
         """Clean the test environment."""
 
-        del self.ocb_key, self.pysat_key, self.notes, self.isvector
-        del self.added_keys, self.pysat_keys, self.arevectors, self.pysat_lat
-        del self.test_file, self.log_capture, self.test_inst, self.ocb
-        del self.lwarn, self.lout, self.cust_kwargs, self.pysat_var2
-        del self.del_time
+        del self.ocb_key, self.pysat_key, self.notes, self.pysat_lat
+        del self.added_keys, self.pysat_keys, self.test_inst, self.ocb
+        del self.test_file, self.log_capture, self.cust_kwargs, self.pysat_var2
+        del self.lwarn, self.lout, self.del_time
         return
 
 
@@ -1452,10 +1388,8 @@ class TestPysatCustMethodsXarray(TestPysatCustMethods):
         self.pysat_key = 'dummy1'
         self.pysat_lat = 'latitude'
         self.notes = None
-        self.isvector = False
         self.added_keys = list()
         self.pysat_keys = list()
-        self.arevectors = list()
         self.del_time = 600
 
         # Set the method defaults
@@ -1492,10 +1426,9 @@ class TestPysatCustMethodsXarray(TestPysatCustMethods):
     def tearDown(self):
         """Clean the test environment."""
 
-        del self.ocb_key, self.pysat_key, self.notes, self.isvector
-        del self.added_keys, self.pysat_keys, self.arevectors
-        del self.test_file, self.log_capture, self.test_inst, self.ocb
-        del self.lwarn, self.lout, self.cust_kwargs, self.pysat_var2
+        del self.ocb_key, self.pysat_key, self.notes, self.lwarn, self.lout
+        del self.added_keys, self.pysat_keys, self.test_inst, self.ocb
+        del self.test_file, self.log_capture, self.cust_kwargs, self.pysat_var2
         del self.del_time, self.pysat_lat
         return
 
@@ -1511,10 +1444,8 @@ class TestPysatCustMethods2DXarray(TestPysatCustMethods):
         self.pysat_key = 'dummy1'
         self.pysat_lat = 'latitude'
         self.notes = None
-        self.isvector = False
         self.added_keys = list()
         self.pysat_keys = list()
-        self.arevectors = list()
         self.del_time = 600
 
         # Set the method defaults
@@ -1547,10 +1478,9 @@ class TestPysatCustMethods2DXarray(TestPysatCustMethods):
     def tearDown(self):
         """Clean the test environment."""
 
-        del self.ocb_key, self.pysat_key, self.notes, self.isvector
-        del self.added_keys, self.pysat_keys, self.arevectors
-        del self.test_file, self.log_capture, self.test_inst, self.ocb
-        del self.lwarn, self.lout, self.cust_kwargs, self.pysat_var2
+        del self.ocb_key, self.pysat_key, self.notes, self.lwarn, self.lout
+        del self.added_keys, self.pysat_keys, self.test_inst, self.ocb
+        del self.test_file, self.log_capture, self.cust_kwargs, self.pysat_var2
         del self.del_time, self.pysat_lat
         return
 
@@ -1564,12 +1494,10 @@ class TestPysatCustMethodsModel(TestPysatCustMethods):
         # Set the utility defaults
         self.ocb_key = 'ocb_test'
         self.pysat_key = 'dummy2'
-        self.pysat_lat = 'mlat'
+        self.pysat_lat = 'latitude'
         self.notes = None
-        self.isvector = False
         self.added_keys = list()
         self.pysat_keys = list()
-        self.arevectors = list()
         self.del_time = 600
 
         # Set the method defaults
@@ -1581,15 +1509,12 @@ class TestPysatCustMethodsModel(TestPysatCustMethods):
         self.ocb.rec_ind = 27
 
         self.test_inst = pysat.Instrument('pysat', 'testmodel',
-                                          num_samples=50400,
                                           clean_level='clean',
                                           update_files=True,
                                           file_date_range=pds.date_range(
                                               self.ocb.dtime[0],
                                               self.ocb.dtime[-1], freq='1D'))
         self.pysat_var2 = 'dummy2'
-        self.test_inst.custom_attach(add_mag_coords,
-                                     kwargs={"dat": self.pysat_var2})
         self.cust_kwargs = {'mlat_name': self.pysat_lat, 'mlt_name': 'mlt',
                             'ocb': self.ocb, 'max_sdiff': self.del_time}
 
@@ -1604,9 +1529,8 @@ class TestPysatCustMethodsModel(TestPysatCustMethods):
     def tearDown(self):
         """Clean the test environment."""
 
-        del self.ocb_key, self.pysat_key, self.notes, self.isvector
-        del self.added_keys, self.pysat_keys, self.arevectors
-        del self.test_file, self.log_capture, self.test_inst, self.ocb
-        del self.lwarn, self.lout, self.cust_kwargs, self.pysat_var2
+        del self.ocb_key, self.pysat_key, self.notes, self.lwarn, self.lout
+        del self.added_keys, self.pysat_keys, self.test_inst, self.ocb
+        del self.test_file, self.log_capture, self.cust_kwargs, self.pysat_var2
         del self.del_time, self.pysat_lat
         return